Refactored markdown-to-html and sanitize-html
