--- web/views/social.pug
@@ -1,4 +1,0 @@
-extends ./includes/layout
-
-block content
-  h1.secundaryColor Coming soon
